+++ src/bw/entrypoint.sh
@@ -13,6 +13,8 @@ log "ENTRYPOINT" "ℹ️" "Starting BunkerWeb v$(cat /usr/share/bunkerweb/VERSIO
 # setup and check /data folder
 /usr/share/bunkerweb/helpers/data.sh "ENTRYPOINT"
 
+handle_docker_secrets
+
 # trap SIGTERM and SIGINT
 function trap_exit() {
 	# shellcheck disable=SC2317